Change Name
BoundCheckPointPtrArray -> BoundCheckAccountPtrArray
AccountPointer -> ACCOUNT_PTR
