+++ bankapp_g/bankapp_g/bankapp_g.vcxproj
@@ -140,7 +140,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="AccountHandler.cpp" />
-    <ClCompile Include="BoundCheckPointPtrArray.cpp" />
+    <ClCompile Include="BoundCheckAccountPtrArray.cpp" />
     <ClCompile Include="HighCreditAccount.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="Account.cpp" />
@@ -149,7 +149,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="AccountHandler.h" />
-    <ClInclude Include="BoundCheckPointPtrArray.h" />
+    <ClInclude Include="BoundCheckAccountPtrArray.h" />
     <ClInclude Include="HighCreditAccount.h" />
     <ClInclude Include="NormalAccount.h" />
   </ItemGroup>
